project/python/telegram_reader/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
distro==1.9.0
docstring-parser==0.15
fastapi==0.115.12
frozenlist==1.5.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
instructor==0.6.8
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.3.1
mypy==1.15.0
mypy-extensions==1.0.0
openai==1.70.0
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
packaging==24.2
pluggy==1.5.0
propcache==0.3.1
protobuf==5.29.4
psycopg2==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
python-json-logger==3.3.0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rsa==4.9
shell_gpt==1.4.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
Telethon==1.39.0
tenacity==8.5.0
tqdm==4.67.1
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
